+++ heimdallm/bifrosts/sql/sqlite/tests/sql/select/test_conditions.py
@@ -1,4 +1,4 @@
-from typing import Sequence
+from typing import Sequence, Type
 
 import pytest
 
@@ -6,6 +6,7 @@ from heimdallm.bifrosts.sql import exc
 from heimdallm.bifrosts.sql.common import FqColumn, JoinCondition
 from heimdallm.bifrosts.sql.sqlite.select.bifrost import Bifrost
 
+from ..utils import dialects
 from .utils import PermissiveConstraints
 
 
@@ -17,13 +18,14 @@ class MyConstraints(PermissiveConstraints):
         return column.name in {"t1.name", "t1.thing"}
 
 
-def test_allowed_where():
+@dialects()
+def test_allowed_where(Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked(MyConstraints())
 
     query = """
     select t1.col from t1
     where
-        t1.name='foo'
+        t1.`name`='foo'
         and t1.thing='bar'
         or t1.other='baz'
     """
@@ -35,13 +37,14 @@ def test_allowed_where():
     query = """
     select t1.col from t1
     where
-        t1.name='foo'
+        t1.`name`='foo'
         and t1.thing='bar'
     """
     bifrost.traverse(query)
 
 
-def test_allowed_join():
+@dialects()
+def test_allowed_join(Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked(MyConstraints())
 
     query = """
@@ -60,12 +63,13 @@ def test_allowed_join():
     bifrost.traverse(query)
 
 
-def test_allowed_having():
+@dialects()
+def test_allowed_having(Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked(MyConstraints())
 
     query = """
-    select t1.name as name, sum(t1.amount) total from t1
-    group by name
+    select t1.`name` as `name`, sum(t1.amount) total from t1
+    group by `name`
     having total > 100
     """
     with pytest.raises(exc.IllegalConditionColumn) as e:
@@ -73,14 +77,15 @@ def test_allowed_having():
     assert e.value.column.name == "t1.amount"
 
     query = """
-    select t1.name as name, sum(t1.thing) total from t1
-    group by name
+    select t1.`name` as `name`, sum(t1.thing) total from t1
+    group by `name`
     having total > 100
     """
     bifrost.traverse(query)
 
 
-def test_allowed_order_by():
+@dialects()
+def test_allowed_order_by(Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked(MyConstraints())
 
     query = """
@@ -93,6 +98,6 @@ def test_allowed_order_by():
 
     query = """
     select t1.id from t1
-    order by t1.name
+    order by t1.`name`
     """
     bifrost.traverse(query)