+++ heimdallm/bifrosts/sql/sqlite/tests/sql/select/test_select.py
@@ -1,4 +1,4 @@
-from typing import Sequence
+from typing import Sequence, Type
 
 import pytest
 
@@ -6,10 +6,12 @@ from heimdallm.bifrosts.sql import exc
 from heimdallm.bifrosts.sql.common import FqColumn, JoinCondition, RequiredConstraint
 from heimdallm.bifrosts.sql.sqlite.select.bifrost import Bifrost
 
+from ..utils import dialects
 from .utils import PermissiveConstraints
 
 
-def test_aliased_select_column():
+@dialects()
+def test_aliased_select_column(Bifrost: Type[Bifrost]):
     """tests that we can select a column from a table that was aliased via a
     join"""
     query = """
@@ -34,7 +36,8 @@ def test_unqualified_columns():
     assert e.value.column == "col"
 
 
-def test_unqualified_where_column():
+@dialects()
+def test_unqualified_where_column(Bifrost: Type[Bifrost]):
     """a column alias in the where class is valid, as long as the alias points to an
     actual aliased column. if it doesn't, we consider it unqualified."""
     query = """
@@ -51,6 +54,7 @@ LIMIT 5;
     assert e.value.column == "customer_id"
 
 
+@dialects("sqlite")
 @pytest.mark.parametrize(
     "query",
     (
@@ -60,7 +64,7 @@ LIMIT 5;
         'select t1.col from t1 join "t2" on t1.jid = "t2".jid',
     ),
 )
-def test_escapes(query):
+def test_escapes(Bifrost: Type[Bifrost], query):
     class MyConstraints(PermissiveConstraints):
         def select_column_allowed(self, column: FqColumn):
             return column.name in {"t1.col"}
@@ -72,7 +76,8 @@ def test_escapes(query):
     bifrost.traverse(query)
 
 
-def test_disallowed_select_column():
+@dialects()
+def test_disallowed_select_column(Bifrost: Type[Bifrost]):
     """tests that our constraint validator works when trying to select a
     column"""
     query = "select t1.col from t1"
@@ -96,7 +101,8 @@ def test_disallowed_select_column():
     assert e.column == "t1.col"
 
 
-def test_required_constraint():
+@dialects()
+def test_required_constraint(Bifrost: Type[Bifrost]):
     """tests that our required constraint restricts the query correctly"""
 
     # missing constraint
@@ -134,6 +140,7 @@ def test_required_constraint():
         bifrost.traverse(query)
 
 
+@dialects()
 @pytest.mark.parametrize(
     "query",
     (
@@ -141,7 +148,7 @@ def test_required_constraint():
         "select `` fro t1",  # just broken
     ),
 )
-def test_broken_query(query):
+def test_broken_query(Bifrost: Type[Bifrost], query):
     """tests that we raise an exception when we cannot parse the query"""
 
     bifrost = Bifrost.mocked(PermissiveConstraints())
@@ -149,12 +156,14 @@ def test_broken_query(query):
         bifrost.traverse(query)
 
 
-def test_select_column_arith():
+@dialects()
+def test_select_column_arith(Bifrost: Type[Bifrost]):
     query = "select t1.col + 1 as plus_one from t1"
     bifrost = Bifrost.mocked(PermissiveConstraints())
     bifrost.traverse(query)
 
 
+@dialects()
 @pytest.mark.parametrize(
     "query",
     (
@@ -162,13 +171,14 @@ def test_select_column_arith():
         "select t1.id, (1 + 1) as two from t1",
     ),
 )
-def test_select_expr(query):
+def test_select_expr(Bifrost: Type[Bifrost], query):
     """select a non-column expression"""
     bifrost = Bifrost.mocked(PermissiveConstraints())
     bifrost.traverse(query)
 
 
-def test_conflicting_validations():
+@dialects()
+def test_conflicting_validations(Bifrost: Type[Bifrost]):
     query = "select t1.col from t1"
 
     class MyConstraints(PermissiveConstraints):
@@ -182,7 +192,8 @@ def test_conflicting_validations():
     bifrost.traverse(query)
 
 
-def test_count_disallowed_column():
+@dialects()
+def test_count_disallowed_column(Bifrost: Type[Bifrost]):
     """we're allowed to count a disallowed column"""
 
     class GeneralConstraints(PermissiveConstraints):