+++ heimdallm/bifrosts/sql/sqlite/tests/sql/select/test_validators.py
@@ -1,9 +1,12 @@
+from typing import Type
+
 import pytest
 
 from heimdallm.bifrosts.sql import exc
 from heimdallm.bifrosts.sql.common import FqColumn
 from heimdallm.bifrosts.sql.sqlite.select.bifrost import Bifrost
 
+from ..utils import dialects
 from .utils import PermissiveConstraints
 
 
@@ -27,50 +30,54 @@ class Succeed2(PermissiveConstraints):
         return True
 
 
-def test_all_fail():
+@dialects()
+def test_all_fail(Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked([Fail1(), Fail2()])
 
     query = """
     select t1.col from t1
     where
-        t1.name='foo'
+        t1.foo='bar'
     """
     with pytest.raises(exc.IllegalConditionColumn) as e:
         bifrost.traverse(query)
-    assert e.value.column.name == "t1.name"
+    assert e.value.column.name == "t1.foo"
 
 
-def test_all_fail_order():
+@dialects()
+def test_all_fail_order(Bifrost: Type[Bifrost]):
     """we always raise the last error, switching the validators proves this."""
     bifrost = Bifrost.mocked([Fail2(), Fail1()])
 
     query = """
     select t1.col from t1
     where
-        t1.name='foo'
+        t1.foo='bar'
     """
     with pytest.raises(exc.IllegalSelectedColumn) as e:
         bifrost.traverse(query)
     assert e.value.column == "t1.col"
 
 
-def test_one_succeed():
+@dialects()
+def test_one_succeed(Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked([Fail1(), Succeed1()])
 
     query = """
     select t1.col from t1
     where
-        t1.name='foo'
+        t1.foo='bar'
     """
     bifrost.traverse(query)
 
 
-def test_both_succeed():
+@dialects()
+def test_both_succeed(Bifrost: Type[Bifrost]):
     bifrost = Bifrost.mocked([Succeed1(), Succeed2()])
 
     query = """
     select t1.col from t1
     where
-        t1.name='foo'
+        t1.foo='bar'
     """
     bifrost.traverse(query)