+++ heimdallm/bifrosts/sql/sqlite/tests/sql/utils.py
@@ -1,15 +1,38 @@
 import pytest
 
 
-def dialects(*dialects):
+def dialects(*dialects, bifrost=True, envelope=False):
+    dialects = dialects or ("sqlite", "mysql")
+
     def map_dialect(d):
-        mod = __import__(f"heimdallm.bifrosts.sql.{d}.select.bifrost")
-        bifrost = getattr(mod.bifrosts.sql, d).select.bifrost.Bifrost
-        bifrost.dialect = d
-        return bifrost
+        args = []
+
+        if bifrost:
+            bf_mod = __import__(f"heimdallm.bifrosts.sql.{d}.select.bifrost")
+            bifrost_cls = getattr(bf_mod.bifrosts.sql, d).select.bifrost.Bifrost
+            bifrost_cls.dialect = d
+            args.append(bifrost_cls)
+
+        if envelope:
+            env_mod = __import__(f"heimdallm.bifrosts.sql.{d}.select.envelope")
+            envelope_cls = getattr(
+                env_mod.bifrosts.sql, d
+            ).select.envelope.PromptEnvelope
+            envelope_cls.dialect = d
+            args.append(envelope_cls)
+
+        if len(args) == 1:
+            return args[0]
+        return args
+
+    args = []
+    if bifrost:
+        args.append("Bifrost")
+    if envelope:
+        args.append("PromptEnvelope")
 
     return pytest.mark.parametrize(
-        "Bifrost",
+        ",".join(args),
         [map_dialect(d) for d in dialects],
         ids=dialects,
     )