+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Bugfix where disabling reconstruction resulted in an error
 - Allow counting disallowed columns
 - Bugfix api key not set when using openai completion method
+- Generalize tests to handle multiple sql dialects
 
 ## 0.1.4 - 7/5/23